Rename opttoy2 to v2
