+++ v2/logic_test.go
@@ -1,4 +1,4 @@
-package opttoy
+package v2
 
 import (
 	"bufio"
