+++ v2/opttoy.go
@@ -1,4 +1,4 @@
-package opttoy
+package v2
 
 import (
 	"bytes"
